Check whether being an UI test matters
